requirements.txt
requests>=2.4.2
six
python-gitlab
pygithub
